# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accessify API — convert academic content into accessible multimodal formats
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SourceType = (typeof SourceType)[keyof typeof SourceType];

export const SourceType = {
  pdf: "pdf",
  image: "image",
  text: "text",
  audio: "audio",
} as const;

export type ReadingLevel = (typeof ReadingLevel)[keyof typeof ReadingLevel];

export const ReadingLevel = {
  elementary: "elementary",
  middle: "middle",
  high: "high",
  original: "original",
} as const;

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  processing: "processing",
  ready: "ready",
  failed: "failed",
} as const;

export type AccessibilityIssueSeverity =
  (typeof AccessibilityIssueSeverity)[keyof typeof AccessibilityIssueSeverity];

export const AccessibilityIssueSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export interface AccessibilityIssue {
  id: string;
  severity: AccessibilityIssueSeverity;
  category: string;
  message: string;
  suggestion: string;
  fixed: boolean;
}

export interface StructuredChunk {
  index: number;
  /** @nullable */
  heading?: string | null;
  text: string;
  /** @nullable */
  signLanguageGloss?: string | null;
  startMs: number;
  endMs: number;
}

export interface DocumentSummary {
  id: number;
  title: string;
  sourceType: SourceType;
  status: DocumentStatus;
  accessibilityScore: number;
  issueCount: number;
  hasAudio: boolean;
  hasSimplified: boolean;
  hasCaptions: boolean;
  createdAt: string;
}

export type DocumentKeyTermsItem = {
  term: string;
  definition: string;
};

export interface Document {
  id: number;
  title: string;
  sourceType: SourceType;
  status: DocumentStatus;
  originalText: string;
  /** @nullable */
  extractedText?: string | null;
  /** @nullable */
  simplifiedText?: string | null;
  readingLevel: ReadingLevel;
  /** @nullable */
  summary?: string | null;
  keyTerms: DocumentKeyTermsItem[];
  /** @nullable */
  audioDataUrl?: string | null;
  /** @nullable */
  audioVoice?: string | null;
  captions: StructuredChunk[];
  accessibilityScore: number;
  issues: AccessibilityIssue[];
  /** @nullable */
  altText?: string | null;
  shareToken: string;
  createdAt: string;
  updatedAt: string;
}

export type SharedDocumentKeyTermsItem = {
  term: string;
  definition: string;
};

export interface SharedDocument {
  title: string;
  sourceType: SourceType;
  readingLevel: ReadingLevel;
  readingText: string;
  /** @nullable */
  summary?: string | null;
  keyTerms: SharedDocumentKeyTermsItem[];
  /** @nullable */
  audioDataUrl?: string | null;
  captions: StructuredChunk[];
}

export type StatsDocumentsBySourceTypeItem = {
  sourceType: SourceType;
  count: number;
};

export type StatsScoreDistributionItem = {
  bucket: string;
  count: number;
};

export interface Stats {
  totalDocuments: number;
  documentsThisWeek: number;
  averageAccessibilityScore: number;
  totalIssuesFixed: number;
  documentsBySourceType: StatsDocumentsBySourceTypeItem[];
  scoreDistribution: StatsScoreDistributionItem[];
}

export interface ActivityItem {
  id: number;
  documentId: number;
  documentTitle: string;
  action: string;
  message: string;
  createdAt: string;
}

export type CreateDocumentBody = {
  /** @minLength 1 */
  title: string;
  sourceType: SourceType;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  imageDataUrl?: string | null;
  /** @nullable */
  audioDataUrl?: string | null;
};

export type SimplifyDocumentBody = {
  readingLevel: ReadingLevel;
};

export type GenerateDocumentAudioBodyVoice =
  (typeof GenerateDocumentAudioBodyVoice)[keyof typeof GenerateDocumentAudioBodyVoice];

export const GenerateDocumentAudioBodyVoice = {
  alloy: "alloy",
  echo: "echo",
  fable: "fable",
  onyx: "onyx",
  nova: "nova",
  shimmer: "shimmer",
} as const;

export type GenerateDocumentAudioBody = {
  voice: GenerateDocumentAudioBodyVoice;
  useSimplified: boolean;
};

export type TranscribeDocumentAudioBody = {
  audioDataUrl: string;
};