# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Accessify API — convert academic content into accessible multimodal formats
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Document upload and accessibility transformation
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSummary"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload a new document and process it for accessibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                sourceType:
                  $ref: "#/components/schemas/SourceType"
                text:
                  type: ["string", "null"]
                imageDataUrl:
                  type: ["string", "null"]
                audioDataUrl:
                  type: ["string", "null"]
              required:
                - title
                - sourceType
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document with all accessible formats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /documents/{id}/simplify:
    post:
      operationId: simplifyDocument
      tags: [documents]
      summary: Generate or regenerate a simplified version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                readingLevel:
                  $ref: "#/components/schemas/ReadingLevel"
              required:
                - readingLevel
      responses:
        "200":
          description: Simplified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}/audio:
    post:
      operationId: generateDocumentAudio
      tags: [documents]
      summary: Generate text-to-speech audio for the document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                voice:
                  type: string
                  enum: [alloy, echo, fable, onyx, nova, shimmer]
                useSimplified:
                  type: boolean
              required: [voice, useSimplified]
      responses:
        "200":
          description: Audio generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}/captions:
    post:
      operationId: generateDocumentCaptions
      tags: [documents]
      summary: Generate timed captions and structured chunks (sign-language ready)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Captions generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}/analyze:
    post:
      operationId: analyzeDocument
      tags: [documents]
      summary: Faculty — analyze accessibility gaps and recommend fixes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}/auto-fix:
    post:
      operationId: autoFixDocument
      tags: [documents]
      summary: Faculty — apply automated accessibility improvements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Improvements applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}/transcribe:
    post:
      operationId: transcribeDocumentAudio
      tags: [documents]
      summary: Transcribe an uploaded audio/lecture file (ASR)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                audioDataUrl:
                  type: string
              required: [audioDataUrl]
      responses:
        "200":
          description: Transcribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /shared/{token}:
    get:
      operationId: getSharedDocument
      tags: [documents]
      summary: Public read-only view of a document via share token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedDocument"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Dashboard statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Recent activity feed
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SourceType:
      type: string
      enum: [pdf, image, text, audio]

    ReadingLevel:
      type: string
      enum: [elementary, middle, high, original]

    DocumentStatus:
      type: string
      enum: [processing, ready, failed]

    AccessibilityIssue:
      type: object
      properties:
        id:
          type: string
        severity:
          type: string
          enum: [critical, warning, info]
        category:
          type: string
        message:
          type: string
        suggestion:
          type: string
        fixed:
          type: boolean
      required: [id, severity, category, message, suggestion, fixed]

    StructuredChunk:
      type: object
      properties:
        index:
          type: integer
        heading:
          type: ["string", "null"]
        text:
          type: string
        signLanguageGloss:
          type: ["string", "null"]
        startMs:
          type: integer
        endMs:
          type: integer
      required: [index, text, startMs, endMs]

    DocumentSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sourceType:
          $ref: "#/components/schemas/SourceType"
        status:
          $ref: "#/components/schemas/DocumentStatus"
        accessibilityScore:
          type: integer
        issueCount:
          type: integer
        hasAudio:
          type: boolean
        hasSimplified:
          type: boolean
        hasCaptions:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, sourceType, status, accessibilityScore, issueCount, hasAudio, hasSimplified, hasCaptions, createdAt]

    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sourceType:
          $ref: "#/components/schemas/SourceType"
        status:
          $ref: "#/components/schemas/DocumentStatus"
        originalText:
          type: string
        extractedText:
          type: ["string", "null"]
        simplifiedText:
          type: ["string", "null"]
        readingLevel:
          $ref: "#/components/schemas/ReadingLevel"
        summary:
          type: ["string", "null"]
        keyTerms:
          type: array
          items:
            type: object
            properties:
              term:
                type: string
              definition:
                type: string
            required: [term, definition]
        audioDataUrl:
          type: ["string", "null"]
        audioVoice:
          type: ["string", "null"]
        captions:
          type: array
          items:
            $ref: "#/components/schemas/StructuredChunk"
        accessibilityScore:
          type: integer
        issues:
          type: array
          items:
            $ref: "#/components/schemas/AccessibilityIssue"
        altText:
          type: ["string", "null"]
        shareToken:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - sourceType
        - status
        - originalText
        - readingLevel
        - keyTerms
        - captions
        - accessibilityScore
        - issues
        - shareToken
        - createdAt
        - updatedAt

    SharedDocument:
      type: object
      properties:
        title:
          type: string
        sourceType:
          $ref: "#/components/schemas/SourceType"
        readingLevel:
          $ref: "#/components/schemas/ReadingLevel"
        readingText:
          type: string
        summary:
          type: ["string", "null"]
        keyTerms:
          type: array
          items:
            type: object
            properties:
              term:
                type: string
              definition:
                type: string
            required: [term, definition]
        audioDataUrl:
          type: ["string", "null"]
        captions:
          type: array
          items:
            $ref: "#/components/schemas/StructuredChunk"
      required:
        - title
        - sourceType
        - readingLevel
        - readingText
        - keyTerms
        - captions

    Stats:
      type: object
      properties:
        totalDocuments:
          type: integer
        documentsThisWeek:
          type: integer
        averageAccessibilityScore:
          type: integer
        totalIssuesFixed:
          type: integer
        documentsBySourceType:
          type: array
          items:
            type: object
            properties:
              sourceType:
                $ref: "#/components/schemas/SourceType"
              count:
                type: integer
            required: [sourceType, count]
        scoreDistribution:
          type: array
          items:
            type: object
            properties:
              bucket:
                type: string
              count:
                type: integer
            required: [bucket, count]
      required:
        - totalDocuments
        - documentsThisWeek
        - averageAccessibilityScore
        - totalIssuesFixed
        - documentsBySourceType
        - scoreDistribution

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        documentTitle:
          type: string
        action:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, documentId, documentTitle, action, message, createdAt]