# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accessify API — convert academic content into accessible multimodal formats
 * OpenAPI spec version: 0.1.0
 */
import type { AccessibilityIssue } from "./accessibilityIssue";
import type { DocumentKeyTermsItem } from "./documentKeyTermsItem";
import type { DocumentStatus } from "./documentStatus";
import type { ReadingLevel } from "./readingLevel";
import type { SourceType } from "./sourceType";
import type { StructuredChunk } from "./structuredChunk";

export interface Document {
  id: number;
  title: string;
  sourceType: SourceType;
  status: DocumentStatus;
  originalText: string;
  /** @nullable */
  extractedText?: string | null;
  /** @nullable */
  simplifiedText?: string | null;
  readingLevel: ReadingLevel;
  /** @nullable */
  summary?: string | null;
  keyTerms: DocumentKeyTermsItem[];
  /** @nullable */
  audioDataUrl?: string | null;
  /** @nullable */
  audioVoice?: string | null;
  captions: StructuredChunk[];
  accessibilityScore: number;
  issues: AccessibilityIssue[];
  /** @nullable */
  altText?: string | null;
  shareToken: string;
  createdAt: Date;
  updatedAt: Date;
}