# lib/api-zod/src/generated/types/sharedDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accessify API — convert academic content into accessible multimodal formats
 * OpenAPI spec version: 0.1.0
 */
import type { ReadingLevel } from "./readingLevel";
import type { SharedDocumentKeyTermsItem } from "./sharedDocumentKeyTermsItem";
import type { SourceType } from "./sourceType";
import type { StructuredChunk } from "./structuredChunk";

export interface SharedDocument {
  title: string;
  sourceType: SourceType;
  readingLevel: ReadingLevel;
  readingText: string;
  /** @nullable */
  summary?: string | null;
  keyTerms: SharedDocumentKeyTermsItem[];
  /** @nullable */
  audioDataUrl?: string | null;
  captions: StructuredChunk[];
}